move ff-specific geometry methods to dedicated file in form_factor/geometry.py
